Refactor: encodedUrl 생성 시 PK를 이용한 방식에서 외부의 유니크 키값을 받아 사용하는 방식으로 변경
+ 인코딩 로직 도메인에서 애플리케이션 레이어로 이동
