--- src/main/java/com/young/shortenerurl/application/dto/UrlCreateRequest.java
@@ -1,6 +1,0 @@
-package com.young.shortenerurl.application.dto;
-
-import com.young.shortenerurl.model.EncodingType;
-
-public record UrlCreateRequest(String originUrl, EncodingType encodingType) {
-}
