+++ src/main/java/com/young/shortenerurl/global/generator/Sequence.java
@@ -1,0 +1,21 @@
+package com.young.shortenerurl.global.generator;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "sequence")
+@Getter
+@NoArgsConstructor
+public class Sequence {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+}
