+++ src/main/java/com/young/shortenerurl/global/generator/SequenceRepository.java
@@ -1,0 +1,6 @@
+package com.young.shortenerurl.global.generator;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SequenceRepository extends JpaRepository<Sequence, Long> {
+}
