+++ src/main/java/com/young/shortenerurl/global/generator/UniqueKeyGenerator.java
@@ -1,0 +1,6 @@
+package com.young.shortenerurl.global.generator;
+
+public interface UniqueKeyGenerator {
+
+    Long generateKey();
+}
