+++ src/main/java/com/young/shortenerurl/url/application/dto/UrlCreateRequest.java
@@ -1,0 +1,6 @@
+package com.young.shortenerurl.url.application.dto;
+
+import com.young.shortenerurl.url.model.EncodingType;
+
+public record UrlCreateRequest(String originUrl, EncodingType encodingType) {
+}
