+++ src/main/java/com/young/shortenerurl/url/infrastructures/UrlJpaRepository.java
@@ -1,6 +1,6 @@
-package com.young.shortenerurl.infrastructures;
+package com.young.shortenerurl.url.infrastructures;
 
-import com.young.shortenerurl.model.Url;
+import com.young.shortenerurl.url.model.Url;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
@@ -12,4 +12,5 @@ public interface UrlJpaRepository extends JpaRepository<Url, Long> {
     @Query("SELECT u FROM Url u WHERE u.encodedUrl.encodedUrl = :encodedUrl")
     Optional<Url> findByEncodedUrl(@Param(value = "encodedUrl") String encodedUrl);
 
+    Optional<Url> findUrlByOriginUrl(String originUrl);
 }
