+++ src/main/java/com/young/shortenerurl/url/presentation/dto/UrlCreateApiRequest.java
@@ -1,6 +1,6 @@
-package com.young.shortenerurl.presentation.dto;
+package com.young.shortenerurl.url.presentation.dto;
 
-import com.young.shortenerurl.model.EncodingType;
+import com.young.shortenerurl.url.model.EncodingType;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
