+++ src/main/resources/templates/index.html
@@ -36,8 +36,8 @@
         <form id="url-form">
             <div>
                 <select class="form-select" name="encodingType" aria-label="Default select example" >
-                    <option value="BASE_62">Base62 Encoding</option>
-                    <option value="URL_SAFE_BASE_64">Url Safe Base64 Encoding</option>
+                    <option value="BASE_64_V2">Base64 V2 Encoding</option>
+                    <option value="BASE_64_V1">Base64 V1 Encoding</option>
                 </select>
             </div>
             <div>
